+++ NEWS.md
@@ -1,5 +1,12 @@
 # Version 1.3.0.99
 
+* `melt.data.frame()` gains a `factorsAsStrings` argument (`TRUE` by default)
+  that controls whether factors are converted to character when melted as
+  measure variables.
+
+* `melt.data.frame()` preserves identical attributes for measure variables,
+  and now throws a warning if they are dropped.
+
 * `melt.data.frame()` gains an internal Rcpp / C++ implementation, and
   is now many orders of magnitudes faster. (Thanks to Kevin Ushey)